+++ wlsh_nodes.py
@@ -704,7 +704,7 @@ class WLSH_Image_Scale_By_Factor:
 
     @classmethod
     def INPUT_TYPES(s):
-        return {"required": { "original": ("IMAGE",), "upscaled": ("IMAGE",),
+        return {"required": { "original": ("IMAGE",),
                               "upscale_method": (s.upscale_methods,),
                               "factor": ("FLOAT", {"default": 2.0, "min": 0.1, "max": 8.0, "step": 0.1})
                               }}
@@ -713,16 +713,17 @@ class WLSH_Image_Scale_By_Factor:
 
     CATEGORY = "WLSH Nodes/upscaling"
 
-    def upscale(self, original, upscaled, upscale_method, factor, crop):
+    def upscale(self, original, upscale_method, factor):
         old_width = original.shape[2]
         old_height = original.shape[1]
         new_width= int(old_width * factor)
         new_height = int(old_height * factor)
         print("Processing image with shape: ",old_width,"x",old_height,"to ",new_width,"x",new_height)
-        samples = upscaled.movedim(-1,1)
+        samples = original.movedim(-1,1)
         s = comfy.utils.common_upscale(samples, new_width, new_height, upscale_method, crop="disabled")
         s = s.movedim(1,-1)
         return (s,)
+
 
 class WLSH_SDXL_Quick_Image_Scale:
     upscale_methods = ["nearest-exact", "bilinear", "area"]