+++ changelog.md
@@ -4,6 +4,8 @@
 
 - Added image preprocessing node to make images grayscale
 
+- Fixed "Image Upscale by Factor" node logic
+
 ## 2023-09-30
 
 - Added non-INFO variant of file IO nodes to quick-fix a compatibility issue that seems to crop up with the TTN suite